Theoretical code
Shooter state machine: Auto detect spike in RPM to detect that an artifact has been shot

Spindex/Intake state machine: Created this so there is a 0.2 second delay from when the beam break detects artifact and the spindex start to move.

Started accounting for number of artifacts inside the robot.